% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonus.R
\name{bonus}
\alias{bonus}
\title{Bonus Question}
\usage{
bonus(x, search_func)
}
\arguments{
\item{x}{The word used to create funnel words.}

\item{search_func}{A function to check to see if a candidate word is in the
a dictionary.}
}
\description{
Create a list of all funnel words that exist in a dictionary.
}
