% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonus2.R
\name{bonus2}
\alias{bonus2}
\title{Bonus 2}
\usage{
bonus2(search_func)
}
\arguments{
\item{search_func}{A function to check to see if a candidate word is in the
a dictionary.}
}
\description{
Find all words in the \code{enable1} dictionary that contains 5 legal funnel words.
}
