% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/funnel.R
\name{cpp_funnel}
\alias{cpp_funnel}
\alias{funnel}
\alias{r_funnel}
\title{Word funnel}
\usage{
cpp_funnel(a, b)

r_funnel(x, y)
}
\arguments{
\item{x}{The base word}

\item{y}{The candidate word}
}
\value{
A logical indicating if \code{x} can go into \code{y}
}
\description{
Word funnel
}
