% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnelslice.R
\name{funnelslice}
\alias{funnelslice}
\title{Funnel and slice plot generator - wrapped for export to shiny}
\usage{
funnelslice(x, plot_title = "NY Cardiac Surgery", plot = "funnel",
  xlabel = "Number of operations per hospital",
  ylabel = "Survival rate (\%)", rank = "precision", riskadj = F,
  RASRplot = F, mean.target = T, plot.target = F, ypercent = T,
  tails = c(0.001, 0.025))
}
\arguments{
\item{x}{A data frame of mortality observations from a number of hospitals.}

\item{plot_title}{The plot target}

\item{plot}{The plot type}

\item{xlabel}{x axis label}

\item{ylabel}{y axis label}

\item{rank}{The plot order - by "precision" or alphabetically by "name"}

\item{riskadj}{Make a risk adjusted plot if TRUE}

\item{RASRplot}{Work with risk adjusted mortality data if TRUE, else raw mortality rates}

\item{mean.target}{Calculate target from mean}

\item{plot.target}{Plot target if TRUE}

\item{ypercent}{Show survival as percentage if true}

\item{tails}{A pair of p-values determining the inner and outer confidence limits}
}
\description{
\code{funnelslice} This function wraps the core funnel4 function,
reducing the parameters to those used by the shinyfunnelapp package.
}
\seealso{
\code{\link{funnel4}}
}

