% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pt2_bonus2}
\alias{pt2_bonus2}
\title{Part 2 Bonus 2}
\usage{
pt2_bonus2(x, wordset, depth = 1L)
}
\arguments{
\item{x}{The word to test.}

\item{wordset}{The list of valid words as an external pointer to a C++
unordered set.}
}
\description{
Find the depth of a word funnel when more than one letter can be skipped in
a single step.
}
