% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_func}
\alias{search_func}
\alias{sequential_search}
\alias{binary_search}
\alias{set_search}
\title{Search Functions}
\usage{
sequential_search(table)

binary_search(table)

set_search(table)
}
\arguments{
\item{table}{A character vector of dictionary words.}
}
\description{
Create a closure that will search a dictionary for the presence of a word.
}
